+++ R/utils-metadata.R
@@ -13,6 +13,7 @@ fill_metadata_template <- function(meta) {
   if (endsWith(local_meta$url, "/")) {
     local_meta$url <- paste0(local_meta$url, "index.html")
   }
+  local_meta$lang <- sub("_", "-", local_meta$lang %||% "en")
   title <- local_meta$pagetitle
   if (grepl("<", title, fixed = TRUE)) {
     local_meta$pagetitle <- xml2::xml_text(xml2::read_html(title))